Align templates and images

Align templates and images

align(
  stimuli,
  pt1 = 0,
  pt2 = 1,
  x1 = NULL,
  y1 = NULL,
  x2 = NULL,
  y2 = NULL,
  width = NULL,
  height = NULL,
  ref_img = 1,
  fill = wm_opts("fill"),
  patch = FALSE,
  squash = FALSE,
  procrustes = FALSE
)

Arguments

stimuli
list of class stimlist

pt1
The first point to align (defaults to 0)

pt2
The second point to align (defaults to 1)

x1
The x-coordinate to align the first point to

y1
The y-coordinate to align the first point to

x2
The x-coordinate to align the second point to

y2
The y-coordinate to align the second point to

width
The width of the aligned image

height
The height of the aligned image

ref_img
The reference image (by index or name) to get coordinates and dimensions from if NULL (defaults to first image)

fill
background color if cropping goes outside the original image

patch
whether to use the patch function to set the background color

squash
whether to move template points outside the image boundaries inside the image

procrustes
Whether to do a procrustes alignment

Value

stimlist with aligned tems and/or images

Examples

stimuli <- demo_stim("lisa")[1:2]

# align to bottom-centre of nose
one_pt <- align(stimuli, pt1 = 55, pt2 = 55)

# align to pupils of second image
two_pt <- align(stimuli, ref_img = 2)

# procrustes align
proc <- align(stimuli, procrustes = TRUE)