Face++ Auto-Delineation

Automatically delineate faces using dlib in python or Face++ (an external service). Since each delineation counts against a daily limit, you need to set up your own Face++ account (see details below).

auto_delin(
  stimuli,
  style = c("dlib7", "dlib70", "fpp106", "fpp83"),
  replace = FALSE,
  face = 1,
  dlib_path = NULL
)

Arguments

stimuli

list of class stimlist

style

Which service and number of landmarks (dlib7, dlib70, fpp106, fpp83)

replace

if FALSE, only gets templates for images with no template

face

which face to delineate in each image if there is more than 1

dlib_path

path to a custom dlib.dat landmark file to use (style is ignored if set)

Value

stimlist with templates

Details

To use Face++ auto-delineation, you need to get your own free API key from https://www.faceplusplus.com. After signing up for an account, go to https://console.faceplusplus.com/app/apikey/list and request a free API key. Add the key and secret to your .Renviron file as follows:

FACEPLUSPLUS_KEY="1234567890abcdefghijk"

FACEPLUSPLUS_SECRET="1234567890abcdefghijk"

Examples

if (FALSE) {
  # requires python
  auto_high <- demo_stim()[1] |>
    auto_delin(replace = TRUE) # replace existing templates

  # requires an API key in .Renviron
  auto_low <- demo_stim()[1] |>
    auto_delin(style = "fpp106", replace = TRUE)

  c(auto_high, auto_low) |> draw_tem() |> plot()
}