Get Point Coordinates

Get a data frame of the x and y coordinates of a template point

Usage

get_point(stimuli, pt = 0)

Arguments

stimuli

list of stimuli with templates

pt

point(s) to return

Value

data frame of x and y coordinates of the specified point(s) for each stimulus

See also

Examples

demo_stim() |> get_point(0:1)
#>     image point        x        y
#> 1 f_multi     0 210.7407 226.4815
#> 2 f_multi     1 288.1482 226.3889
#> 3 m_multi     0 211.1111 214.4444
#> 4 m_multi     1 286.4815 215.5556