Label with magick annotations

This is just a wrapper function for magick::image_annotate that allows more flexibility in color input. Setting a font, weight, style only works if your imagemagick is compiled with fontconfig support.

mlabel(
  stimuli,
  text = TRUE,
  gravity = "north",
  location = "+0+0",
  degrees = 0,
  size = 0.1,
  font = "sans",
  style = "normal",
  weight = 400,
  kerning = 0,
  decoration = NULL,
  color = "black",
  strokecolor = NULL,
  boxcolor = NULL
)

Arguments

stimuli

list of class stimlist

text

a vector of the label text(s) or TRUE to use stimlist names

gravity

string with gravity value from magick::gravity_types.

location

geometry string with location relative to gravity

degrees

rotates text around center point

size

font size in pixels or proportion of image width (if < 1.0)

font

string with font family such as "sans", "mono", "serif", "Times", "Helvetica", "Trebuchet", "Georgia", "Palatino" or "Comic Sans".

style

value of style_types for example "italic"

weight

thickness of the font, 400 is normal and 700 is bold.

kerning

increases or decreases whitespace between letters

decoration

value of decoration_types: "LineThrough" "None", "Overline", "Underline"

color

a vector of the label colour(s)

strokecolor

adds a stroke (border around the text)

boxcolor

adds a background color

Value

stimlist with labelled images

See also

Examples

stimuli <- demo_stim("test")
labelled_stimuli <- stimuli |>
  mlabel(text = c("CHINWE", "GEORGE"), 
        gravity = c("north", "south"),
        color = "red")